Refactor trait name to QueriesFlaggedEnums
